Refactor access token handling
The method that the API was using to get an access token was failing sometimes because the access token is expired or other problems. To resolve that would have needed user interaction to refresh the access token or even to log in again.

This commit adds a new context to handle the access token and a provider to wrap the app with it. If there is a problem with the token, it provides a flag to the view code that triggers a button to be displayed for the user to click and refresh the token.

Upshot: any code can depend on having an up-to-date access token.